backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-dotenv==1.0.1
requests==2.32.4
Pillow==11.0.0

# 보안 관련
django-ratelimit==4.1.0
django-extensions==3.2.3

# 개발 도구
ipython==8.29.0

# 프로덕션 의존성 (선택사항)
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2 